+++ breeze-core/src/main/kotlin/extension/CharExtensions.kt
@@ -10,8 +10,12 @@ package com.windea.breezeframework.core.extension
  * 重复当前字符到指定次数。
  * @see com.windea.breezeframework.core.extension.repeat
  */
-operator fun Char.times(n: Int): String = this.repeat(n)
+operator fun Char.times(n: Int): String {
+	return this.repeat(n)
+}
 
 
 /**重复当前字符到指定次数。*/
-fun Char.repeat(n: Int): String = String(CharArray(n) { this })
+fun Char.repeat(n: Int): String {
+	return String(CharArray(n) { this })
+}