+++ breeze-core/src/main/kotlin/extension/ComponentExtensions.kt
@@ -1,6 +1,8 @@
 // Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
 // Breeze is blowing...
 
+@file:JvmName("ComponentExtensions")
+
 package com.windea.breezeframework.core.extension
 
 import com.windea.breezeframework.core.annotation.*