+++ breeze-core/src/main/kotlin/extension/IteratorExtensions.kt
@@ -5,9 +5,12 @@
 
 package com.windea.breezeframework.core.extension
 
+import com.windea.breezeframework.core.annotation.*
+
 /**
  * 遍历迭代器中的元素到指定条件处，或者返回null。
  */
+@UnstableApi
 fun <T> Iterator<T>.next(predicate: (current: T,prev:T?) -> Boolean): T? {
 	var current:T
 	var prev:T? = null